FE API: logic for articles fetching via Elasticsearch is now moved to the frontend-api package
