+++ packages/frontend-api/src/Resources/config/graphql-types/ArticleSiteDecorator.types.yaml
@@ -1,0 +1,24 @@
+ArticleSiteDecorator:
+    type: object
+    decorator: true
+    inherits:
+        - 'ArticleDecorator'
+    config:
+        interfaces:
+            - 'Breadcrumb'
+            - 'Slug'
+            - 'NotBlogArticleInterface'
+            - 'ArticleInterface'
+        fields:
+            createdAt:
+                type: "DateTime!"
+                description: "Date and time of the article creation"
+            external:
+                type: "Boolean!"
+                description: "If the the article should be open in a new tab"
+            slug:
+                type: "String!"
+                description: "Article URL slug"
+            breadcrumb:
+                type: "[Link!]!"
+                description: "Hierarchy of the current element in relation to the structure"
