+++ packages/frontend-api/src/Resources/config/graphql-types/QueryDecorator.types.yaml
@@ -88,15 +88,27 @@ QueryDecorator:
                         type: "String"
                 description: "Returns order filtered using UUID or urlHash"
             articles:
-                type: "ArticleConnection"
+                type: "ArticleConnection!"
                 argsBuilder: "Relay::Connection"
                 resolve: "@=query('articlesQuery', args, args['placement'])"
+                complexity: "@=dynamicPaginationComplexity(args)"
                 args:
                     placement:
-                        type: "String"
+                        type: "[ArticlePlacementTypeEnum!]"
+                        defaultValue: []
+                        validation:
+                            -   NotNull: ~
+                        description: "An array of the required articles placements"
                 description: "Returns list of articles that can be paginated using `first`, `last`, `before` and `after` keywords and filtered by `placement`"
+            articlesSearch:
+                type: "[ArticleInterface!]!"
+                resolve: "@=query('articlesSearchQuery', args)"
+                args:
+                    search:
+                        type: "String!"
+                description: "Returns list of searched articles and blog articles"
             article:
-                type: 'Article'
+                type: 'NotBlogArticleInterface'
                 resolve: "@=query('articleByUuidOrUrlSlugQuery', args['uuid'], args['urlSlug'])"
                 args:
                     uuid:
@@ -105,17 +117,21 @@ QueryDecorator:
                         type: "String"
                 description: "Returns article filtered using UUID or URL slug"
             termsAndConditionsArticle:
-                type: 'Article'
+                type: 'ArticleSite'
                 resolve: "@=query('termsAndConditionsArticleQuery')"
                 description: "Returns Terms and Conditions article"
             privacyPolicyArticle:
-                type: 'Article'
+                type: 'ArticleSite'
                 resolve: "@=query('privacyPolicyArticleQuery')"
                 description: "Returns privacy policy article"
             cookiesArticle:
-                type: 'Article'
+                type: 'ArticleSite'
                 resolve: "@=query('cookiesArticleQuery')"
                 description: "Returns information about cookies article"
+            ArticleSite:
+                type: "ArticleSite"
+            ArticleLink:
+                type: "ArticleLink"
             brands:
                 type: '[Brand!]!'
                 resolve: "@=query('brandsQuery')"
