+++ project-base/app/config/graphql/types/ModelType/Article/ArticleLink.types.yaml
@@ -1,24 +1,4 @@
 ArticleLink:
     type: object
-    config:
-        interfaces:
-            - 'NotBlogArticleInterface'
-        fields:
-            uuid:
-                type: "Uuid!"
-                description: "UUID of the article link"
-            placement:
-                type: "String!"
-                description: "Placement of the article link"
-            name:
-                type: "String!"
-                description: "Name of article link, used as anchor text"
-            createdAt:
-                type: "DateTime!"
-                description: "Creation date time of the article link"
-            url:
-                type: "String!"
-                description: "Destination url of article link"
-            external:
-                type: "Boolean!"
-                description: "If the the article should be open in a new tab"
+    inherits:
+        - 'ArticleLinkDecorator'
