+++ project-base/app/config/graphql/types/ModelType/Article/ArticleSite.types.yaml
@@ -1,23 +1,4 @@
 ArticleSite:
     type: object
     inherits:
-        - 'ArticleDecorator'
-    config:
-        interfaces:
-            - 'Breadcrumb'
-            - 'Slug'
-            - 'NotBlogArticleInterface'
-            - 'ArticleInterface'
-        fields:
-            createdAt:
-                type: "DateTime!"
-                description: "Date and time of the article creation"
-            external:
-                type: "Boolean!"
-                description: "If the the article should be open in a new tab"
-            slug:
-                type: "String!"
-                description: "Article URL slug"
-            breadcrumb:
-                type: "[Link!]!"
-                description: "Hierarchy of the current element in relation to the structure"
+        - 'ArticleSiteDecorator'
