--- project-base/app/config/graphql/types/Query/Article/ArticleQuery.types.yaml
@@ -1,49 +1,0 @@
-ArticleQuery:
-    type: object
-    config:
-        fields:
-            articles:
-                type: "ArticleConnection!"
-                argsBuilder: "Relay::Connection"
-                resolve: "@=query('articlesQuery', args, args['placement'])"
-                complexity: "@=dynamicPaginationComplexity(args)"
-                args:
-                    placement:
-                        type: "[ArticlePlacementTypeEnum!]"
-                        defaultValue: []
-                        validation:
-                            -   NotNull: ~
-                        description: "An array of the required articles placements"
-                description: "Returns list of articles that can be paginated using `first`, `last`, `before` and `after` keywords and filtered by `placement`"
-            articlesSearch:
-                type: "[ArticleInterface!]!"
-                resolve: "@=query('articlesSearchQuery', args)"
-                args:
-                    search:
-                        type: "String!"
-                description: "Returns list of searched articles and blog articles"
-            article:
-                type: 'NotBlogArticleInterface'
-                resolve: "@=query('articleQuery', args['uuid'], args['urlSlug'])"
-                args:
-                    uuid:
-                        type: "Uuid"
-                    urlSlug:
-                        type: "String"
-                description: "Returns article filtered using UUID or URL slug"
-            termsAndConditionsArticle:
-                type: 'ArticleSite'
-                resolve: "@=query('termsAndConditionsArticleQuery')"
-                description: "Returns Terms and Conditions article"
-            privacyPolicyArticle:
-                type: 'ArticleSite'
-                resolve: "@=query('privacyPolicyArticleQuery')"
-                description: "Returns privacy policy article"
-            cookiesArticle:
-                type: 'ArticleSite'
-                resolve: "@=query('cookiesArticleQuery')"
-                description: "Returns information about cookies article"
-            ArticleSite:
-                type: "ArticleSite"
-            ArticleLink:
-                type: "ArticleLink"
