+++ project-base/app/config/graphql/types/Query/Query.types.yaml
@@ -7,7 +7,6 @@ Query:
         - 'CategoryQuery'
         - 'StoreQuery'
         - 'SliderQuery'
-        - 'ArticleQuery'
         - 'FlagQuery'
         - 'OrderQuery'
         - 'PersonalDataQuery'
