+++ project-base/app/config/services_frontend_api.yaml
@@ -76,14 +76,6 @@ services:
         alias: App\FrontendApi\Resolver\Category\CategoryQuery
         public: true
 
-    Shopsys\FrontendApiBundle\Model\Resolver\Article\ArticleQuery:
-        alias: App\FrontendApi\Resolver\Article\ArticleQuery
-        public: true
-
-    Shopsys\FrontendApiBundle\Model\Resolver\Article\ArticlesQuery:
-        alias: App\FrontendApi\Resolver\Article\ArticlesQuery
-        public: true
-
     Shopsys\FrontendApiBundle\Model\Mutation\Order\CreateOrderMutation:
         alias: App\FrontendApi\Mutation\Order\CreateOrderMutation
